Version update
Dashboard is updated to version 1.1.5 from 1.0.0
The download link is therefore set to the latest version. This update includes stable serialization fixes of course contents, notification-dating problems, tasks, etc. News feeds can now be read as closely as they are on the browser. Swing components like JEditorPane particularly helps achieve this convenience. There is also a change in the Transcript file-naming, as might be an issue during testing, or mass transcript exportation. This is in 60-seconds precision. Finally, there is an update security mechanism for the Collapse & Launch Sequences. Due to the tremendous requests demanded by these sequences in a limited duration, the native system may potentially ignore some requests, causing Dashboard to unnecessarily cross the base-directory. This is particularly common in Windows Systems.